fix and addition
fixed a bug: one-frame movies weren't exporting, because the framework was assigning the frame a duration of 0 :: added an option for specifying a fallback framerate ("AVFallbackFPSKey"), which is used to set the duration of the frame when exporting one-frame movies.  if the fallback framerate hasn't been specified, the duration of the frame will be set to the minimum frame duration.  this corresponds with build 1.6.9.1 of the batch converter.